refactor: move class Keyboard to single file
